backend/shared/requirements.txt
# Python dependencies for RootTrust shared utilities
pydantic==2.5.0
pydantic[email]==2.5.0
boto3==1.34.0
bcrypt==4.1.2
PyJWT==2.8.0
